+++ example/lib/main.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: avoid_print
+
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:open_street_map_flutter/open_street_map_flutter.dart';
@@ -32,6 +34,7 @@ class MyHomePage extends StatefulWidget {
 
 class _MyHomePageState extends State<MyHomePage> {
   ByteData? icon;
+  String cameraState = "";
   String cameraPosition = "";
   OpenStreetMapController? mapController;
 
@@ -45,34 +48,48 @@ class _MyHomePageState extends State<MyHomePage> {
         children: [
           Expanded(
               child: OpenStreetMap(
-            initialCameraPosition:
-                CameraPosition(center: LatLng(52.3532222, 9.7582331), zoom: 13),
+            initialCameraPosition: const CameraPosition(
+                center: LatLng(52.3532222, 9.7582331), zoom: 13),
             enableMyLocation: true,
-            style: Style(invertColors: false),
+            style: const Style(invertColors: false),
             onMapCreated: (controller) {
-              this.mapController = controller;
+              mapController = controller;
+            },
+            onCameraMoveStarted: () {
+              setState(() {
+                print("Moving started..");
+                cameraState = "Moving started";
+              });
             },
             onCameraMove: (CameraPosition cameraPosition) {
               setState(() {
+                print("Moving..");
+                cameraState = "Moving";
                 this.cameraPosition = cameraPosition.toString();
+              });
+            },
+            onCameraIdle: (CameraPosition cameraPosition) {
+              print("Camera idle..");
+              setState(() {
+                cameraState = "Idle";
               });
             },
             markers: {
               Marker(
                   id: "marker-1",
-                  point: LatLng(52.3532222, 9.7582331),
+                  point: const LatLng(52.3532222, 9.7582331),
                   icon: icon,
                   onTap: () {
-                    ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+                    ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
                       content: Text("Clicked marker one."),
                     ));
                   }),
               Marker(
                   id: "marker-2",
-                  point: LatLng(52.346984, 9.7584736),
+                  point: const LatLng(52.346984, 9.7584736),
                   icon: icon,
                   onTap: () {
-                    ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+                    ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
                       content: Text("Clicked marker two."),
                     ));
                   }),
@@ -83,21 +100,22 @@ class _MyHomePageState extends State<MyHomePage> {
                   color: Colors.orange,
                   width: 10,
                   points: {
-                    LatLng(52.34841, 9.75083),
-                    LatLng(52.35434, 9.74628),
-                    LatLng(52.35681, 9.76301)
+                    const LatLng(52.34841, 9.75083),
+                    const LatLng(52.35434, 9.74628),
+                    const LatLng(52.35681, 9.76301)
                   })
             },
           )),
-          Text("$cameraPosition"),
+          Text(cameraState),
+          Text(cameraPosition),
         ],
       ),
       floatingActionButton: FloatingActionButton(
           onPressed: () {
-            mapController?.animateCamera(CameraPosition(
+            mapController?.animateCamera(const CameraPosition(
                 center: LatLng(52.3881111, 9.6944627), zoom: 17.0));
           },
-          child: Icon(Icons.move_down_rounded)),
+          child: const Icon(Icons.move_down_rounded)),
     );
   }
 
@@ -109,7 +127,7 @@ class _MyHomePageState extends State<MyHomePage> {
   }
 
   void _loadIcon() async {
-    String resolutionName = await AssetImage('images/marker.png')
+    String resolutionName = await const AssetImage('images/marker.png')
         .obtainKey(ImageConfiguration.empty)
         .then((value) => value.name);