+++ CHANGELOG.md
@@ -1,3 +1,15 @@
-## 0.0.1
+CHANGELOG
+=========
 
-* TODO: Describe initial release.
+## 0.1.0 ##
+
+* ➕ Set camera
+* ➕ Animate camera
+* ➕ Set markers
+* ➕ Set polylines
+* ➕ Show my location
+* ➕ Invert colors
+* ➕ Android Plugin
+* ➕ iOS Plugin
+
+